[feat]퀘스트 수락

+++ src/main/java/site/FitUp/main/api/stat/dtos/StatResponse.java
@@ -11,7 +11,7 @@ public class StatResponse {
     @Builder
     @Getter
     public static class CreateStatResponse{
-        private int userStatSeq;
+        private long userStatSeq;
         private int strength;
         private int endurance;
         private int speed;
